stowaway can latejoin, and doesn't show a message

+++ code/controllers/subsystems/ticker.dm
@@ -179,7 +179,7 @@ var/global/datum/controller/subsystem/ticker/ticker
 		//Cleanup some stuff
 		for(var/obj/effect/landmark/start/S in landmarks_list)
 			//Deleting Startpoints but we need the ai point to AI-ize people later
-			if (S.name != "AI")
+			if (S.name != "AI" && S.name != "Stowaway")
 				qdel(S)
 		to_world("<span class='boldannounce notice'><em>Enjoy the game!</em></span>")
 		//YW Change start: Custom message thanks to VerySoft and a 5% chance to play a secret message
